Partially migrate server code to TS
