+++ src/js/server/errors.ts
@@ -1,12 +1,12 @@
-const chalk = require('chalk');
+import chalk from 'chalk';
 
-exports.middleware = function (err, req, res, next) {
+export function middleware(err, req, res, next) {
   const message = err.stack || err.toString();
   console.error(chalk.red.underline('Error occurred. We goofed.'));
   console.error(message);
   res.status(err.status || 500);
 
-  const locals = { };
+  const locals: any = { };
 
   if (req.app.get('env') === 'development') {
     locals.message = message;
